requirements.txt
flask
flask-mail
flask-sqlalchemy
flask-login
flask-wtf
flask-script
flask-migrate
psycopg2